+++ SportsMeet/Data/DataBase.cs
@@ -226,7 +226,7 @@ namespace SportsMeet.Data
 
         #region Events
 
-        internal static List<Event> GetEventsForEventIds(List<long> eventIds)
+        internal static List<Event> GetEventsForEventIds(List<int> eventIds)
         {
             List<Event> eventList = new List<Event>();
             foreach (var eventId in eventIds)
@@ -622,10 +622,10 @@ namespace SportsMeet.Data
             return playerList;
         }
 
-        public static List<long> GetPlayerEventsNotByPlayer(PlayerEvent searchPlayerEvents)
+        public static List<int> GetPlayerEventsNotByPlayer(PlayerEvent searchPlayerEvents)
         {
             string query = "select id from Events where id not IN (select eventId from PlayerEvents where playerId = @PlayerId);";
-            IEnumerable<long> result = DBConnection.Instance.Connection.Query<long>(query, searchPlayerEvents);
+            IEnumerable<int> result = DBConnection.Instance.Connection.Query<int>(query, searchPlayerEvents);
             return result.ToList();
         }
